% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_notify.R
\name{current_issue_metadata}
\alias{current_issue_metadata}
\title{Obtain the current R Weekly issue metadata}
\usage{
current_issue_metadata(schedule_df, curator_df, current_date = Sys.Date())
}
\arguments{
\item{schedule_df}{data frame of the R Weekly schedule of curation}

\item{curator_df}{data frame of current R Weekly curators}

\item{current_date}{string with the current date, default is `Sys.Date()`}
}
\value{
list with the following components:
* `current_issue_id`: String of issue identifier in YYYY-WXX format
* `curator_name`: String with full name of scheduled issue curator
* `curator_id`: String with the Slack ID of scheduled issue curator
* `backup_name`: String with the full name of the scheduled issue backup
* `backup_id`: String with the Slack ID of the scheduled issue backup
}
\description{
Obtain the current R Weekly issue metadata
}
