% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_curator_data}
\alias{import_curator_data}
\title{Import R Weekly curator team data}
\usage{
import_curator_data(
  query = "SELECT * FROM curator_df",
  from_owner = "rweekly-org",
  from_repo_name = "curation-schedule"
)
}
\arguments{
\item{query}{String with the SQL querty to execute on Dolthub. Default is
the string `SELECT * FROM curator_df` which simply returns all rows of the 
table.}

\item{from_owner}{String with the Dolthub owner of the database. Default is
`rweekly-org`.}

\item{from_repo_name}{String with the name of the repository containing the
database. Default is `curation-schedule`.}
}
\value{
data frame of the R Weekly curator data
}
\description{
Obtain the current version of the R Weekly curator team metadata
from the R Weekly Dolthub repository
}
