% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_notify.R
\name{reminder_driver}
\alias{reminder_driver}
\title{Driver function to send issue reminder}
\usage{
reminder_driver(channel = "#dev", notify_user = TRUE)
}
\arguments{
\item{channel}{String with the Slack channel for the message. Default is `#dev`.}

\item{notify_user}{Indicator for tagging (notifying) any users mentioned in the 
message. Default is `TRUE`.}
}
\description{
This function is intended to be used in an automated step such as a 
GitHub action. It is simply a wrapper around the functions to obtain
the upcoming issue metadata and sending the notification in Slack.
}
