% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_slack.R
\name{send_reminder_message}
\alias{send_reminder_message}
\title{Send current issue curation reminder in Slack}
\usage{
send_reminder_message(
  current_issue_id,
  curator_id,
  curator_name,
  backup_id,
  backup_name,
  channel = "#random",
  username = "rweeklycalbot",
  notify_user = TRUE
)
}
\arguments{
\item{current_issue_id}{String of issue identifier in YYYY-WXX format}

\item{curator_id}{String with the Slack ID of scheduled issue curator}

\item{curator_name}{String with full name of scheduled issue curator}

\item{backup_id}{String with the Slack ID of the scheduled issue backup}

\item{backup_name}{String with the full name of the scheduled issue backup}

\item{channel}{String with the Slack channel for the message. Default is `#random`.}

\item{username}{String iwth the Slack ID of the bot. Default is `rweeklycalbot`.}

\item{notify_user}{Indicator for tagging (notifying) any users mentioned in the 
message. Default is `TRUE`.}
}
\value{
Nothing, used for the message side effect
}
\description{
Send current issue curation reminder in Slack
}
